+++ nb-engine/nb-engine-core/src/test/java/io/nosqlbench/engine/core/ActivityExecutorTest.java
@@ -30,6 +30,7 @@ import io.nosqlbench.engine.api.activityimpl.motor.CoreMotor;
 import io.nosqlbench.engine.api.activityimpl.motor.CoreMotorDispenser;
 import io.nosqlbench.engine.core.lifecycle.ExecutionResult;
 import io.nosqlbench.engine.core.lifecycle.activity.ActivityExecutor;
+import io.nosqlbench.engine.core.lifecycle.activity.ActivityTypeLoader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.Test;
@@ -54,10 +55,10 @@ class ActivityExecutorTest {
 //        OutputDispenser tdisp = CoreServices.getOutputDispenser(activity).orElse(null);
 //
 //        final MotorDispenser<?> mdisp = new CoreMotorDispenser(activity, inputDispenser, adisp, tdisp);
-//        activity.setActionDispenserDelegate(adisp);
-//        activity.setOutputDispenserDelegate(tdisp);
-//        activity.setInputDispenserDelegate(inputDispenser);
-//        activity.setMotorDispenserDelegate(mdisp);
+//        activity.setActionDispenser(adisp);
+//        activity.setOutputDispenser(tdisp);
+//        activity.setInputDispenser(inputDispenser);
+//        activity.setMotorDispenser(mdisp);
 //
 //        final ExecutorService executor = Executors.newCachedThreadPool();
 //        ActivityExecutor activityExecutor = new ActivityExecutor(activity, "test-restart");
@@ -110,10 +111,10 @@ class ActivityExecutorTest {
         final OutputDispenser outputDispenser = CoreServices.getOutputDispenser(activity).orElse(null);
 
         MotorDispenser<?> motorDispenser = new CoreMotorDispenser(activity, inputDispenser, actionDispenser, outputDispenser);
-        activity.setActionDispenserDelegate(actionDispenser);
-        activity.setOutputDispenserDelegate(outputDispenser);
-        activity.setInputDispenserDelegate(inputDispenser);
-        activity.setMotorDispenserDelegate(motorDispenser);
+        activity.setActionDispenser(actionDispenser);
+        activity.setOutputDispenser(outputDispenser);
+        activity.setInputDispenser(inputDispenser);
+        activity.setMotorDispenser(motorDispenser);
 
         ActivityExecutor activityExecutor = new ActivityExecutor(activity);
 
@@ -138,22 +139,22 @@ class ActivityExecutorTest {
         final ActivityDef activityDef = ActivityDef.parseActivityDef("driver=diag;alias=test_dynamic_params;cycles=1000;initdelay=5000;");
         new ActivityTypeLoader().load(activityDef,TestComponent.INSTANCE);
 
-        Activity simpleActivity = new SimpleActivity(TestComponent.INSTANCE,activityDef);
+        Activity activity = new Activity(TestComponent.INSTANCE, activityDef);
 
-//        this.getActivityMotorFactory(this.motorActionDelay(999), new AtomicInput(simpleActivity,activityDef));
+//        this.getActivityMotorFactory(this.motorActionDelay(999), new AtomicInput(activity,activityDef));
 
-        final InputDispenser inputDispenser = new CoreInputDispenser(simpleActivity);
-        final ActionDispenser actionDispenser = new CoreActionDispenser(simpleActivity);
-        final OutputDispenser outputDispenser = CoreServices.getOutputDispenser(simpleActivity).orElse(null);
+        final InputDispenser inputDispenser = new CoreInputDispenser(activity);
+        final ActionDispenser actionDispenser = new CoreActionDispenser(activity);
+        final OutputDispenser outputDispenser = CoreServices.getOutputDispenser(activity).orElse(null);
 
-        MotorDispenser<?> motorDispenser = new CoreMotorDispenser<>(simpleActivity,
+        MotorDispenser<?> motorDispenser = new CoreMotorDispenser<>(activity,
                 inputDispenser, actionDispenser, outputDispenser);
 
-        simpleActivity.setActionDispenserDelegate(actionDispenser);
-        simpleActivity.setInputDispenserDelegate(inputDispenser);
-        simpleActivity.setMotorDispenserDelegate(motorDispenser);
+        activity.setActionDispenser(actionDispenser);
+        activity.setInputDispenser(inputDispenser);
+        activity.setMotorDispenser(motorDispenser);
 
-        ActivityExecutor activityExecutor = new ActivityExecutor(simpleActivity);
+        ActivityExecutor activityExecutor = new ActivityExecutor(activity);
         activityDef.setThreads(5);
         ForkJoinTask<ExecutionResult> executionResultForkJoinTask = ForkJoinPool.commonPool().submit(activityExecutor);
 
@@ -188,7 +189,7 @@ class ActivityExecutorTest {
         return new MotorDispenser<>() {
             @Override
             public Motor getMotor(final ActivityDef activityDef, final int slotId) {
-                final Activity activity = new SimpleActivity(TestComponent.INSTANCE,activityDef);
+                final Activity activity = new Activity(TestComponent.INSTANCE, activityDef);
                 final Motor<?> cm = new CoreMotor<>(activity, slotId, ls);
                 cm.setAction(lc);
                 return cm;
